backend/requirements.txt
fastapi[standard]==0.117.1
uvicorn==0.37.0
pydantic==2.10.3
python-dotenv==1.1.1
PyMuPDF==1.26.3
requests==2.32.5